+++ client/src/test/spec/controller/AgendaControllerTest.js
@@ -7,20 +7,24 @@
         var self = this;
         const ID_LOCAL_TEST = 1;
 
-        let createController, scope, config, LocaisMock, ReservasMock,
-            AuthService, dependenciasController, $mdDialog;
+        let createController, $scope, config, LocaisMock, ReservasMock,
+            AuthService, dependenciasController, $mdDialog, ModalService,
+            Reserva, TIPOS_RESERVA;
 
         beforeEach(inject(defaultInjections(self)));
         afterEach(defaultAfterEach(self));
 
-        beforeEach(inject(function ($injector, $controller, uiCalendarConfig, _LocaisMock_, _AuthService_, _ReservasMock_) {
+        beforeEach(inject(function ($injector, $controller, uiCalendarConfig, _LocaisMock_, _AuthService_, _ReservasMock_, _ModalService_, _Reserva_, _TIPOS_RESERVA_) {
 
             $mdDialog = $injector.get('$mdDialog');
-            scope = self.$rootScope.$new();
+            $scope = self.$rootScope.$new();
             config = uiCalendarConfig;
             LocaisMock = _LocaisMock_;
             AuthService = _AuthService_;
             ReservasMock = _ReservasMock_;
+            ModalService = _ModalService_;
+            Reserva = _Reserva_;
+            TIPOS_RESERVA = _TIPOS_RESERVA_;
 
             config.calendars = {
                 calendario: {
@@ -31,10 +35,13 @@
             };
 
             dependenciasController = {
-                $scope: scope,
+                $scope,
                 local: LocaisMock.getLocal({_id: ID_LOCAL_TEST}),
                 uiCalendarConfig: config,
-                AuthService: AuthService
+                AuthService,
+                ModalService,
+                Reserva,
+                TIPOS_RESERVA
             };
 
             createController = function () {
@@ -46,7 +53,27 @@
             };
         }));
 
-        describe('calendarioConfig deve', function () {
+        describe('AgendaController visualizarReserva deve', () => {
+            let controller;
+            beforeEach(() => {
+                controller = createController();
+            });
+
+            it('chamar modal de para visualização de reserva e atualizar as reservas no calendario', done => {
+                ModalService.verReserva = sinon.stub().returns(self.$q.when());
+                config.calendars.calendario.fullCalendar = sinon.stub();
+
+                controller.visualizarReserva({}).should.be.fulfilled.then(() => {
+                    sinon.assert.called(ModalService.verReserva);
+
+                    sinon.assert.calledWith(config.calendars.calendario.fullCalendar, 'removeEvents');
+                    sinon.assert.calledWith(config.calendars.calendario.fullCalendar, 'addEventSource', _.first(controller.reservasSource));
+                }).should.notify(done);
+                $scope.$digest();
+            });
+        });
+
+        describe('AgendaController calendarioConfig deve', function () {
             let controller;
             beforeEach(() => {
                 controller = createController();
@@ -69,7 +96,7 @@
                 controller = createController();
 
                 AuthService.userTemPermissao = sinon.stub().returns(true);
-                $mdDialog.show = sinon.stub();
+                $mdDialog.show = sinon.stub().returns(self.$q.when());
             }));
 
            it('Não permitir criação de reservas em horários passados', function () {