+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Refatoração: Traduzidos textos dos botões de entrar e sair da aplicação.
 * Refatoração: Mudado local para designar título à reserva.
 * Manutenção: Adicionado limites para horário de início e fim do calendário de acordo com o horário de funcionamento do local.
+* Issue 71: Corrigido nome de "Auditório" para "Local" na visualização da agenda de um local.
+* Issue 60: Corrigida atualização de reservas na agenda.
 * Manutenção: Horário de fim da reserva agora inicia preenchido no modal de criação da mesma.
 
 ## 0.2.0